requirements.txt
crewai
flask
python-dotenv
openai-whisper
Flask-SocketIO